Allow decimal mark-up rates in temp-to-perm calculator
Before this commit a markup rate of '30.2%' would be truncated to
30% (0.3).

+++ app/controllers/temp_to_perm_calculator/home_controller.rb
@@ -11,7 +11,7 @@ module TempToPermCalculator
         days_per_week: params[:days_per_week].to_i,
         contract_start_date: contract_start_date,
         hire_date: hire_date,
-        markup_rate: params[:markup_rate].to_i / 100.0
+        markup_rate: params[:markup_rate].to_f / 100.0
       )
     end
